+++ PS4DebugPage.cs
@@ -1013,8 +1013,8 @@ namespace Dobby {
             BackFunc();
             HoverLeave(BackBtn, false); // What Did This Fix, Again?
         }
-        private void InfoHelpBtn_Click(object sender, EventArgs e) => ChangeForm((int)PageID.InfoHelpPageId);
-        private void CreditsBtn_Click(object sender, EventArgs e) => ChangeForm((int)PageID.CreditsPageId);
+        private void InfoHelpBtn_Click(object sender, EventArgs e) => ChangeForm(PageID.InfoHelpPageId);
+        private void CreditsBtn_Click(object sender, EventArgs e) => ChangeForm(PageID.CreditsPageId);
 
         public void IgnoreTitleIDBtn_Click(object sender, EventArgs e) {
             TempStringStore = ((Control)sender).Text;
